refactor(domain): move ScanOptions + SessionMutationSnapshot into voom-domain (#378)
ScanOptions (with its closure type aliases ErrorCallback / FingerprintLookup /
MutationSnapshotLoader) and SessionMutationSnapshot move from
plugins/discovery into voom-domain so Call::ScanLibrary can carry a typed
options field. The discovery crate re-exports both, leaving existing
voom_discovery::ScanOptions / voom_discovery::SessionMutationSnapshot paths
intact. Defaults (recursive: true, hash_files: true, workers: 0) preserved
verbatim per the rev-3 regression-guard test.
