+++ crates/voom-domain/src/lib.rs
@@ -7,6 +7,7 @@ pub mod capability_resolution;
 pub mod compiled;
 pub mod evaluation;
 pub mod orchestration;
+pub mod scan;
 pub mod errors;
 pub mod estimate;
 pub mod events;
